Rename lib > src
