+++ embedded/xsens_imu/src/i2c.h
