+++ embedded/xsens_imu/src/imu.h
